Hide Open with PlayTorch deep link if were not likely on an android or ios device
