+++ website/src/components/ExpoSnackLandingPage.js
@@ -12,20 +12,7 @@ import clsx from 'clsx';
 import ExpoSnack from '@site/src/components/ExpoSnack';
 import styles from './ExpoSnackLandingPage.module.css';
 import RedirectStarterSnack from './RedirectStarterSnack';
-import ExternalLinks from '@site/src/constants/ExternalLinks';
-import AppleLogoSvg from '@site/static/img/apple_logo_white.svg';
-import AndroidLogoSvg from '@site/static/img/android_logo_white.svg';
-
-function AppStoreButton({href, LogoComponent, label}) {
-  return (
-    <a className="button button--primary margin--xs" href={href}>
-      <div className={clsx(styles.appStoreButtonContainer)}>
-        <LogoComponent className={clsx(styles.appStoreLogo)} />
-        <div className={clsx(styles.appStoreButtonText)}>{label}</div>
-      </div>
-    </a>
-  );
-}
+import LandingPageHeader from './LandingPageHeader';
 
 export default function ExpoSnackLandingPage({match}) {
   const {expoSnackPath} = match.params;
@@ -36,42 +23,7 @@ export default function ExpoSnackLandingPage({match}) {
 
   return (
     <div>
-      <div className="row">
-        <div className="col col--10 col--offset--1">
-          <div className="hero">
-            <div className="container">
-              <h1 className="hero__title">PlayTorch</h1>
-              <p className="hero__subtitle">
-                PlayTorch is a framework for rapidly creating cross-platform
-                mobile AI experiences.
-              </p>
-              <div>
-                <p className="margin-bottom--sm">Already have the app?</p>
-                <a
-                  className="button button--primary margin--xs"
-                  href={`playtorch://snack?snackUrl=exp://exp.host/${expoSnackPath}`}>
-                  Open with PlayTorch
-                </a>
-              </div>
-              <div className="margin-top--lg">
-                <p className="margin-bottom--sm">
-                  This snack requires the PlayTorch app to get started.
-                </p>
-                <AppStoreButton
-                  href={ExternalLinks.APP_STORE}
-                  LogoComponent={AppleLogoSvg}
-                  label="Join the iOS Beta"
-                />
-                <AppStoreButton
-                  href={ExternalLinks.GOOGLE_PLAY_STORE}
-                  LogoComponent={AndroidLogoSvg}
-                  label="Download PlayTorch for Android"
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <LandingPageHeader expoSnackPath={expoSnackPath} />
       <div className={clsx(styles.expoSnackContainer)}>
         <ExpoSnack snackPreview={false} snackId={expoSnackPath} />
       </div>
